Convert Travis CI config to GitHub Actions
Travis CI dependency installation and tests are ported as-is. Travis CI had an
unspecified execution environment; in GitHub Actions, we will run on available
Ubuntu (18.04 and 20.04) and macOS (10.15 and 11) versions.

Additionally Travis CI config has not been updated in a while and only ran with
Python 2.7, 3.3, and 3.4; for GitHub Actions, this commit adds all newer Python
versions (up to 3.10) as well.

Update README.md to use the new GitHub Actions badge and point to the new test
status, which will become active as soon as this PR goes through the testing
process the first time.

Closes https://github.com/rossant/ipycache/issues/52