Explicitly install `nose` to unbreak tests
The command `nosetests` appears to have disappeared from the GitHub Actions
images for Python 3.8 and 3.9 on all versions of Ubuntu and macOS, so let's try
installing it manually to fix the tests in the short-term.

During the development of [this change][1], we've learned that ipycache is not
compatible with recently-released IPython 8, so we updated `requirements.txt` to
require a version of IPython prior to 8; additionally, we re-enabled testing
with Python 2.7 since we'll need to update our code in the future to be
compatible with `nose2`, so until we drop support for Python 2.7, we should be
testing with that version as well.

Replace import `IPython.nbformat.current` with `nbformat` and install `nbformat`
module.

Replace `from IPython.kernel import KernelManager` with
`from jupyter_client.manager import KernelManager` and install `jupyter-client`.

Also install `ipykernel` package to address the error message
"jupyter_client.kernelspec.NoSuchKernel: No such kernel named python3".

Create a separate `test_requirements_python27.txt` with old versions of
dependencies to support Python 2.7, and upgrade dependencies in
`test_requirements.txt` to the latest versions for the newest Python 3.x
releases.

Closes https://github.com/rossant/ipycache/issues/63

[1]: https://github.com/rossant/ipycache/pull/64