Rewrite unit tests to use Python `unittest`
This lets us drop the dependency on the deprecated `nose` library and use the
Python standard library `unittest`. Since we don't use any advanced
functionality, and the `unittest` library is sufficient for our use cases, this
means we can remove the dependency on `nose` without introducing a new
dependency on either `nose2` or `pytest`.

Drop test dependency on `nose` since we no longer need it.

Change the test runner command from `nosetests` to `python -m unittest`.

Closes https://github.com/rossant/ipycache/issues/61